Move Tags assigner to a separate file
